core/requirements.txt
Django==5.2.5
pypdf==5.1.0
python-dotenv==1.0.1
requests==2.32.3
pytesseract==0.3.13
pdf2image==1.17.0
Pillow==11.0.0
channels==4.3.2
daphne==4.2.1
